@@ -1,13 +1,13 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.13'
+gem 'rails', '4.2.0'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
 
 gem 'sqlite3'
 gem 'mysql2'
-gem 'strong_parameters'
+# gem 'strong_parameters'
 gem 'will_paginate', '3.0.7'
 gem 'carrierwave'	, 	'0.10.0'
 gem 'mini_magick'	, 	'3.8.0' 
@@ -15,18 +15,24 @@ gem 'devise'
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+	# Use SCSS for stylesheets
+	gem 'sass-rails', '~> 5.0'
+	# Use Uglifier as compressor for JavaScript assets
+	gem 'uglifier', '>= 1.3.0'
+	# Use CoffeeScript for .coffee assets and views
+	gem 'coffee-rails', '~> 4.1.0'
 
-  # See https://github.com/sstephenson/execjs#readme for more supported runtimes
-  # gem 'therubyracer', :platforms => :ruby
-
-  gem 'uglifier', '>= 1.0.3'
+	#gem 'therubyracer'
+	#gem 'libv8'
+	#gem "less-rails" #Sprockets (what Rails 3.1 uses for its asset pipeline) supports LESS
+	#gem "twitter-bootstrap-rails"
 end
 
 gem 'jquery-rails'
+gem 'turbolinks'
 gem 'active_attr'
 gem "recaptcha", :require => "recaptcha/rails"
+gem "gretel"
 # To use ActiveModel has_secure_password
 # gem 'bcrypt-ruby', '~> 3.0.0'
 
