@@ -1,6 +1,6 @@
 class Article < ActiveRecord::Base
   belongs_to :blog_category, counter_cache: true
-  has_many :comments, as: :commentable
+  has_many :comments, as: :commentable, dependent: :destroy
   validates :title, presence: true
   mount_uploader :home_image, PhotoUploader
   mount_uploader :article_image, PhotoUploader
