@@ -1,4 +1,4 @@
 class BlogCategory < ActiveRecord::Base
-	has_many :articles
+	has_many :articles, dependent: :destroy
   validates :name, presence: true
 end
