@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Miracode::Application.config.secret_token = '173840e29dc300434b48f709599a3141823c0a11fe3350846435da00eab8a167ea207f0844fa22ba95234eb2d95a3528479f9add48c07a4e019a6b1d26c6be25'
+Miracode::Application.config.secret_key_base = '173840e29dc300434b48f709599a3141823c0a11fe3350846435da00eab8a167ea207f0844fa22ba95234eb2d95a3528479f9add48c07a4e019a6b1d26c6be25'
